requirements.txt
django
django-jsonview
